Add (most) missing docs and refactor doc package
