+++ plugin/src/main/kotlin/nl/tudelft/hyperion/plugin/git/GitLineTracker.kt
@@ -8,7 +8,11 @@ import git4idea.commands.Git
 import git4idea.commands.GitCommand
 import git4idea.commands.GitLineHandler
 import git4idea.repo.GitRepository
+import nl.tudelft.hyperion.plugin.git.GitLineTracker.resolveCurrentLine
 
+/**
+ * Class that tracks lines using git. The primary method to use is [resolveCurrentLine].
+ */
 object GitLineTracker {
     /**
      * Takes in a project and file, as well as an old commit hash and a line number. Will
