+++ plugin/src/main/kotlin/nl/tudelft/hyperion/plugin/visualization/MetricInlayRenderPassFactory.kt
@@ -1,4 +1,4 @@
-package nl.tudelft.hyperion.plugin.doc
+package nl.tudelft.hyperion.plugin.visualization
 
 import com.intellij.codeHighlighting.TextEditorHighlightingPass
 import com.intellij.codeHighlighting.TextEditorHighlightingPassFactory
@@ -10,6 +10,9 @@ import com.intellij.openapi.util.Key
 import com.intellij.psi.PsiFile
 import nl.tudelft.hyperion.plugin.util.KeyedProperty
 
+/**
+ * Class that handles the creation of HighlightingPasses [MetricInlayRenderPass] for given file and editor.
+ */
 class MetricInlayRenderPassFactory : TextEditorHighlightingPassFactory, TextEditorHighlightingPassFactoryRegistrar {
     override fun registerHighlightingPassFactory(registrar: TextEditorHighlightingPassRegistrar, project: Project) {
         registrar.registerTextEditorHighlightingPass(this, null, null, false, -1)
@@ -20,7 +23,9 @@ class MetricInlayRenderPassFactory : TextEditorHighlightingPassFactory, TextEdit
 
         // Ignore if the file hasn't changed.
         val savedStamp = editor.modificationStamp
-        val currentStamp = getCurrentModificationStamp(file)
+        val currentStamp = getCurrentModificationStamp(
+            file
+        )
         if (savedStamp != null && savedStamp == currentStamp) return null
 
         return MetricInlayRenderPass(editor, file)
@@ -33,7 +38,9 @@ class MetricInlayRenderPassFactory : TextEditorHighlightingPassFactory, TextEdit
         var Editor.modificationStamp by KeyedProperty(PSI_MODIFICATION_STAMP)
 
         fun putCurrentModificationStamp(editor: Editor, file: PsiFile) {
-            editor.modificationStamp = getCurrentModificationStamp(file)
+            editor.modificationStamp = getCurrentModificationStamp(
+                file
+            )
         }
 
         private fun getCurrentModificationStamp(file: PsiFile): Long {
