+++ plugin/src/main/kotlin/nl/tudelft/hyperion/plugin/visualization/MetricTooltipRenderer.kt
@@ -1,4 +1,4 @@
-package nl.tudelft.hyperion.plugin.doc
+package nl.tudelft.hyperion.plugin.visualization
 
 import com.intellij.codeInsight.daemon.impl.HintRenderer
 import com.intellij.openapi.editor.Inlay
@@ -7,6 +7,11 @@ import com.intellij.openapi.editor.markup.TextAttributes
 import java.awt.Graphics
 import java.awt.Rectangle
 
+/**
+ * Class that represents the Renderer for the [MetricInlayItem].
+ * It is a slightly modified [HintRenderer] that also keeps track of a [RangeHighlighter] to
+ * keep track of the start of the line (excluding tabs and spaces).
+ */
 class MetricTooltipRenderer(
     text: String?,
     private val highlighter: RangeHighlighter
