+++ plugin/src/main/kotlin/nl/tudelft/hyperion/plugin/visualization/RefreshTooltipAction.kt
@@ -3,8 +3,11 @@ package nl.tudelft.hyperion.plugin.visualization
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
-import nl.tudelft.hyperion.plugin.doc.MetricInlayRenderPass
 
+/**
+ * Class that represents the action in the IDE that can be fired by the user.
+ * It resides under the "Tools" menu and has the hotkey CTRL+ALT+H, F
+ */
 class RefreshTooltipAction : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val editor = e.getData(CommonDataKeys.EDITOR)
