+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -63,7 +63,7 @@
         <projectService serviceImplementation="nl.tudelft.hyperion.plugin.settings.HyperionSettings"/>
         <projectConfigurable displayName="Hyperion" groupId="tools"
                              instance="nl.tudelft.hyperion.plugin.settings.HyperionSettingsConfigurable"/>
-        <highlightingPassFactory implementation="nl.tudelft.hyperion.plugin.doc.MetricInlayRenderPassFactory"/>
+        <highlightingPassFactory implementation="nl.tudelft.hyperion.plugin.visualization.MetricInlayRenderPassFactory"/>
         <search.optionContributor implementation="nl.tudelft.hyperion.plugin.settings.HyperionOptionContributor"/>
     </extensions>
 </idea-plugin>
